clean code (#87)
clean code

+++ modules/tools/html/tools_update.php
@@ -34,7 +34,11 @@ if ($update == "UPDATE") {
     include("$ROOT/modules/tools/update_perms.php");
     include("$ROOT/modules/tools/update_db.php");
     include("$ROOT/modules/tools/check_packages.php");
-	unlink("$ROOT/tmp/update");
+	
+	if (file_exists("$ROOT/tmp/update")) {
+		
+            unlink("$ROOT/tmp/update");
+        }
 
     echo '</pre>';
